Refactor API layer to use static mock data for portfolio demo
- Created client/src/api/mockData.ts from db.json
- Updated all API files to import data from mockData.ts
- Removed json-server dependency and script from package.json
- Deleted db.json
- Verified application functionality and login flow with Playwright
- Ensured successful production build with CI=true
