+++ client/package.json
@@ -31,8 +31,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "deploy": "gh-pages -d build",
-    "server": "json-server --watch db.json --port 3001"
+    "deploy": "gh-pages -d build"
   },
   "eslintConfig": {
     "extends": [
@@ -56,7 +55,6 @@
     "@types/node": "^25.0.3",
     "autoprefixer": "^10.4.16",
     "gh-pages": "^6.1.1",
-    "json-server": "^1.0.0-beta.3",
     "postcss": "^8.4.32",
     "prettier": "^3.2.5",
     "prettier-plugin-tailwindcss": "^0.5.11",
