+++ client/src/api/payments.ts
@@ -1,4 +1,4 @@
-import db from '../db.json';
+import { db } from './mockData';
 
 // Use window.location.origin to support any deployment URL (S3, Vercel, Localhost)
 const appUrl = typeof window !== 'undefined' ? window.location.origin : 'http://localhost:3000';
